Refactor julie001.py: extract ProjectXClient and OptimizedTPEngine
Move ProjectXClient class to dedicated client.py module for API
interactions with ProjectX Gateway. Move OptimizedTPEngine class
to risk_engine.py for take profit calculation logic.

julie001.py now focuses strictly on orchestration and the main
execution loop, improving code organization and maintainability.
